+++ internal/gkr/bw6-633/gkr.go
@@ -18,7 +18,7 @@ import (
 	"github.com/consensys/gnark-crypto/utils"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/internal/gkr/gkrtypes"
-	"github.com/consensys/gnark/std/gkrapi/gkr"
+	"github.com/consensys/gnark/std/gkrapi/v2/gkr"
 )
 
 // The goal is to prove/verify evaluations of many instances of the same circuit
