+++ internal/gkr/bw6-761/gkr_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/consensys/gnark/internal/gkr/gkrtesting"
 	"github.com/consensys/gnark/internal/gkr/gkrtypes"
 	"github.com/consensys/gnark/internal/utils"
-	"github.com/consensys/gnark/std/gkrapi/gkr"
+	"github.com/consensys/gnark/std/gkrapi/v2/gkr"
 	"github.com/stretchr/testify/assert"
 )
