+++ std/gkrapi/v2/api_test.go
@@ -1,4 +1,4 @@
-package gkrapi
+package v2
 
 import (
 	"fmt"
@@ -18,7 +18,7 @@ import (
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/frontend/cs/r1cs"
 	"github.com/consensys/gnark/frontend/cs/scs"
-	"github.com/consensys/gnark/std/gkrapi/gkr"
+	"github.com/consensys/gnark/std/gkrapi/v2/gkr"
 	stdHash "github.com/consensys/gnark/std/hash"
 	"github.com/consensys/gnark/test"
 	"github.com/stretchr/testify/assert"
